Adding simple Point class and renaming all var members to remove prefixes.

+++ src/lib/Point.cpp
@@ -1,5 +1,5 @@
 #include "Point.h"
 
 namespace Sax {
-	Point::Point( double x, double y ) : x( x ), y( y ){};
+	Point::Point( int x, int y ) : x( x ), y( y ){};
 }
